fix(menu): keep Technique animations stable over long uptime
Store the shared visual-style background clock as f64 so long main-menu sessions keep sub-frame delta precision instead of eventually quantizing animation updates.

Cache Technique background model geometry with stable textured-mesh keys and keep model animation inputs bounded, avoiding per-frame geometry rebuild churn that could make the Technique main-menu background slow down or appear to stop after extended idle time.

+++ src/screens/init.rs
@@ -755,7 +755,9 @@ fn push_actors_with_elapsed_overrides(
         alpha_mul: 1.0,
     };
     if let Some(bg_elapsed_s) = bg_elapsed_override {
-        state.bg.push_at_elapsed(actors, bg_params, bg_elapsed_s);
+        state
+            .bg
+            .push_at_elapsed(actors, bg_params, f64::from(bg_elapsed_s));
     } else {
         state.bg.push(actors, bg_params);
     }
